chore: move chart code
